% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRejections.R
\name{cell_corr}
\alias{cell_corr}
\title{Cell correlation for simulating data}
\usage{
cell_corr(vname.1, vname.2, rho.XX, rho.YY, rho.XY, nY, prop.corr = 1)
}
\arguments{
\item{vname.1}{Quoted name of first variable}

\item{vname.2}{Quoted name of second variable}

\item{rho.XX}{Correlation between pairs of Xs}

\item{rho.YY}{Correlation between all pairs of Ys}

\item{rho.XY}{rho.XY Correlation between pairs of X-Y (of non-null ones)}

\item{nY}{Number of outcomes}

\item{prop.corr}{Proportion of X-Y pairs that are non-null (non-nulls will be first .prop.corr * .nY pairs)}
}
\description{
The user does not need to call this function. This internal function is called by \code{make_corr_mat}
and populates a single cell. Assumes X1 is the covariate of interest and that none of the covariates is associated with any outcomes.
}
