% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRejections.R
\name{fix_input}
\alias{fix_input}
\title{Fix bad user input}
\usage{
fix_input(X, C, Ys, d)
}
\arguments{
\item{X}{Single quoted name of covariate of interest}

\item{C}{Vector of quoted covariate names}

\item{Ys}{Vector of quoted outcome names}

\item{d}{Dataframe}
}
\description{
The user does not need to call this function. Warns about and fixes bad user input: missing data on analysis variables,
datasets containing extraneous variables, or datasets containing covariates that are not
mean-centered.
}
