% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRejections.R
\name{get_crit}
\alias{get_crit}
\title{Return ordered critical values for Wstep}
\usage{
get_crit(p.dat, col.p)
}
\arguments{
\item{p.dat}{p-values from dataset (W-vector?)}

\item{col.p}{Column of resampled p-values (for the single p-value for which we're}
}
\description{
The user does not need to call this function. This is an internal function for use by 
\code{adj_minP} and \code{adj_Wstep}.
}
